Restructure a folder for semantic segmenation only
