--- src/__init__.py
@@ -1,2 +1,0 @@
-from src.training_function import train
-import src.net_parameters
