--- src/loss/__init__.py
@@ -1,1 +1,0 @@
-from src.loss.dice_loss import DiceCoeff
