+++ src/semantic/__init__.py
@@ -1,0 +1,2 @@
+from src.semantic.training_function import train
+import src.semantic.net_parameters
