+++ src/semantic/loss/__init__.py
@@ -1,0 +1,1 @@
+from src.semantic.loss.dice_loss import DiceCoeff
