+++ src/semantic/net_parameters.py
