+++ src/semantic/unet/__init__.py
@@ -1,0 +1,1 @@
+from src.semantic.unet.generate_masks import create_labels_from_dir
