+++ src/semantic/unet/unet_model.py
@@ -1,5 +1,5 @@
 import torch.nn.functional as F
-from src.unet.unet_utils import *
+from src.semantic.unet.unet_utils import *
 
 class UNet(nn.Module):
     def __init__(self, n_channels, n_classes):
