+++ src/semantic/utils/resize_images.py
