--- src/unet/__init__.py
@@ -1,1 +1,0 @@
-from src.unet.generate_masks import create_labels_from_dir
